+++ src/gpuarray_blas_opencl_clblast.c
@@ -284,8 +284,8 @@ static int ddot(
 
 static int hgemv(cb_order order, cb_transpose transA, size_t M, size_t N,
                  float alpha, gpudata *A, size_t offA, size_t lda,
-                 gpudata *X, size_t offX, size_t incX, float beta,
-                 gpudata *Y, size_t offY, size_t incY) {
+                 gpudata *X, size_t offX, int incX, float beta,
+                 gpudata *Y, size_t offY, int incY) {
   cl_ctx *ctx = A->ctx;
   StatusCode err;
   cl_event ev;
@@ -311,8 +311,8 @@ static int hgemv(cb_order order, cb_transpose transA, size_t M, size_t N,
 
 static int sgemv(cb_order order, cb_transpose transA, size_t M, size_t N,
                  float alpha, gpudata *A, size_t offA, size_t lda,
-                 gpudata *X, size_t offX, size_t incX, float beta,
-                 gpudata *Y, size_t offY, size_t incY) {
+                 gpudata *X, size_t offX, int incX, float beta,
+                 gpudata *Y, size_t offY, int incY) {
   cl_ctx *ctx = A->ctx;
   StatusCode err;
   cl_event ev;
@@ -338,8 +338,8 @@ static int sgemv(cb_order order, cb_transpose transA, size_t M, size_t N,
 
 static int dgemv(cb_order order, cb_transpose transA, size_t M, size_t N,
                  double alpha, gpudata *A, size_t offA, size_t lda,
-                 gpudata *X, size_t offX, size_t incX, double beta,
-                 gpudata *Y, size_t offY, size_t incY) {
+                 gpudata *X, size_t offX, int incX, double beta,
+                 gpudata *Y, size_t offY, int incY) {
   cl_ctx *ctx = A->ctx;
   StatusCode err;
   cl_event ev;
@@ -448,8 +448,8 @@ static int dgemm(cb_order order, cb_transpose transA, cb_transpose transB,
 }
 
 static int hger(cb_order order, size_t M, size_t N, float alpha,
-                gpudata *X, size_t offX, size_t incX,
-                gpudata *Y, size_t offY, size_t incY,
+                gpudata *X, size_t offX, int incX,
+                gpudata *Y, size_t offY, int incY,
                 gpudata *A, size_t offA, size_t lda) {
   cl_ctx *ctx = X->ctx;
   cl_event ev;
@@ -474,8 +474,8 @@ static int hger(cb_order order, size_t M, size_t N, float alpha,
 }
 
 static int sger(cb_order order, size_t M, size_t N, float alpha,
-                gpudata *X, size_t offX, size_t incX,
-                gpudata *Y, size_t offY, size_t incY,
+                gpudata *X, size_t offX, int incX,
+                gpudata *Y, size_t offY, int incY,
                 gpudata *A, size_t offA, size_t lda) {
   cl_ctx *ctx = X->ctx;
   cl_event ev;
@@ -500,8 +500,8 @@ static int sger(cb_order order, size_t M, size_t N, float alpha,
 }
 
 static int dger(cb_order order, size_t M, size_t N, double alpha,
-                gpudata *X, size_t offX, size_t incX,
-                gpudata *Y, size_t offY, size_t incY,
+                gpudata *X, size_t offX, int incX,
+                gpudata *Y, size_t offY, int incY,
                 gpudata *A, size_t offA, size_t lda) {
   cl_ctx *ctx = X->ctx;
   cl_event ev;
